Refactor: relocate interface
